Changed column name in user_procedure_query_history
The name for the timestamp col here should be `logged_timestamp`; not `execution_timestamp`.

Right now I'm thinking its gonna be easiest if we batch the executions, run those, then batch the logs and run that. So the log time and the execution time will be sort of close but not the same. So I don't want to be misleading by calling this time the "execution" time