switching to AWS
I really did try to make this work with Azure. But I've spent so much time on it, I think it's time to move on.\nThis is especially considering the fact that I just had an interview where I'm pretty sure I got dinged for not having AWS Experience.\nSo AWS it is!