stashing ner stuff
possibly im just not using this thing right-- but its getting a bunch of crappy data that isn't very helpul. Just most of the time it doesn't hit anything. And it costs a ton of compute time to run locally, and getting it into batches and back into snowflake isn't cheap either. so scrapping this for now. let's see how topic analysis goes