fixed an interesting dupe-causer
very strange but-- its like this. There is a sproc that inserts rows of data that it transfroms from JSON. sometimes, it creates dupes during insert. I knew this, and added a QUALIFY ROW_NUMBER() ... = 1 and it worked. Then I thought aw thats overengineering let me just make it a DISTINCT. and it stopped working! strangest thing! Even if you do the DISTINCT in a separate CTE it doesnt work. its fixed now but i wonder why this happens...